+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/BraintreeClient.kt
@@ -149,6 +149,8 @@ class BraintreeClient @VisibleForTesting internal constructor(
                 startTime = params.startTime,
                 endTime = params.endTime,
                 endpoint = params.endpoint,
+                experiment = params.experiment,
+                paymentMethodsDisplayed = params.paymentMethodsDisplayed
             )
             sendAnalyticsEvent(event, configuration, authorization)
         }