+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalCheckoutRequestUnitTest.java
@@ -109,7 +109,7 @@ public class PayPalCheckoutRequestUnitTest {
         assertEquals("USD", result.getCurrencyCode());
         assertEquals("en-US", result.getLocaleCode());
         assertEquals("Billing Agreement Description",
-                result.getBillingAgreementDescription());
+            result.getBillingAgreementDescription());
         assertTrue(result.isShippingAddressRequired());
         assertTrue(result.isShippingAddressEditable());
         assertEquals("Postal Address", result.getShippingAddressOverride().getRecipientName());
@@ -210,4 +210,20 @@ public class PayPalCheckoutRequestUnitTest {
         JSONObject jsonObject = new JSONObject(requestBody);
         assertFalse(jsonObject.has("shipping_callback_url"));
     }
+
+    public void createRequestBody_sets_userPhoneNumber_when_not_null() throws JSONException {
+        PayPalCheckoutRequest request = new PayPalCheckoutRequest("1.00", true);
+
+        request.setUserPhoneNumber(new PayPalPhoneNumber("1", "1231231234"));
+        String requestBody = request.createRequestBody(
+            mock(Configuration.class),
+            mock(Authorization.class),
+            "success_url",
+            "cancel_url",
+            null
+        );
+
+        assertTrue(requestBody.contains("\"phone_number\":{\"country_code\":\"1\",\"national_number\":\"1231231234\"}"));
+    }
 }
+